dataengineering/requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
psycopg2==2.9.9
sqlalchemy==2.0.35